move modal in atoms
