+++ src/components/atomic/modals/DavitDraggableModal.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
-import { PositionTO } from "../../../../dataAccess/access/to/PositionTO";
-import { DnDBox, DnDBoxType } from "../../../organisms/dndBox/DnDBox";
+import { PositionTO } from "../../../dataAccess/access/to/PositionTO";
+import { DnDBox, DnDBoxType } from "../../organisms/dndBox/DnDBox";
 import { DavitModal } from "./DavitModal";
 
 interface DavitDraggableModalProps {
