+++ src/components/atomic/modals/DavitModal.css
@@ -1,5 +1,3 @@
-/* -------------------------------------------- Modal -------------------------------------------- */
-
 .davitModal {
     position: fixed;
     top: 0;
