+++ src/components/organisms/controllPanel/presentation/fragments/edit/fragments/modals/EditDecisionModal.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { DavitDraggableModal } from "../../../../../../../common/fragments/modals/DavitDraggableModal";
+import { DavitDraggableModal } from "../../../../../../../atomic/modals/DavitDraggableModal";
 import { DecisionForm } from "../forms/DecisionForm";
 
 interface EditDecicionModalProps {
