+++ src/components/organisms/controllPanel/presentation/fragments/edit/fragments/modals/EditStepModal.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { DavitDraggableModal } from "../../../../../../../common/fragments/modals/DavitDraggableModal";
+import { DavitDraggableModal } from "../../../../../../../atomic/modals/DavitDraggableModal";
 import { StepForm } from "../forms/StepForm";
 
 interface EditStepModalProps {
